Fix: Update `listchars`. Remove `eol` because of an issue
The issue is due to a combination of `alacritty` and the highlighting
group used to represent `eol`. This causes new line characters to be
light grey and when the cursor is over said character it becomes hard
to distinguish the cursor position wrt the background. This can be fixed
by removing the `eol` character thereby making all new lines white
highlighted.